Renamed all files to lowercase to prevent issues in jenkins
