+++ npm-debug.log
@@ -1,0 +1,37 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'run', 'dev-cdn' ]
+2 info using npm@2.14.7
+3 info using node@v4.2.1
+4 verbose run-script [ 'predev-cdn', 'dev-cdn', 'postdev-cdn' ]
+5 info predev-cdn auth0-tutorial-navigator@0.12.9
+6 info dev-cdn auth0-tutorial-navigator@0.12.9
+7 verbose unsafe-perm in lifecycle true
+8 info auth0-tutorial-navigator@0.12.9 Failed to exec dev-cdn script
+9 verbose stack Error: auth0-tutorial-navigator@0.12.9 dev-cdn: `grunt cdn_dev`
+9 verbose stack Exit status 6
+9 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:214:16)
+9 verbose stack     at emitTwo (events.js:87:13)
+9 verbose stack     at EventEmitter.emit (events.js:172:7)
+9 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:24:14)
+9 verbose stack     at emitTwo (events.js:87:13)
+9 verbose stack     at ChildProcess.emit (events.js:172:7)
+9 verbose stack     at maybeClose (internal/child_process.js:818:16)
+9 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:211:5)
+10 verbose pkgid auth0-tutorial-navigator@0.12.9
+11 verbose cwd /Users/Marcos/Repos/tutorial-navigator
+12 error Darwin 15.0.0
+13 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "run" "dev-cdn"
+14 error node v4.2.1
+15 error npm  v2.14.7
+16 error code ELIFECYCLE
+17 error auth0-tutorial-navigator@0.12.9 dev-cdn: `grunt cdn_dev`
+17 error Exit status 6
+18 error Failed at the auth0-tutorial-navigator@0.12.9 dev-cdn script 'grunt cdn_dev'.
+18 error This is most likely a problem with the auth0-tutorial-navigator package,
+18 error not with npm itself.
+18 error Tell the author that this fails on your system:
+18 error     grunt cdn_dev
+18 error You can get their info via:
+18 error     npm owner ls auth0-tutorial-navigator
+18 error There is likely additional logging output above.
+19 verbose exit [ 1, true ]
