+++ src/components/Breadcrumbs.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { getPlatformName, getTechTitle } from '../util/Tutorials';
-import navigateAction from '../action/navigateAction';
-import TutorialStore from '../stores/TutorialStore';
+import { getPlatformName, getTechTitle } from '../util/tutorials';
+import navigateAction from '../action/navigate-action';
+import TutorialStore from '../stores/tutorial-store';
 import { connectToStores, provideContext } from 'fluxible-addons-react';
 
 class Breadcrumbs extends React.Component {
